fix: comments, files division
